+++ apps/studio/src/routes/editor/EditPanel/ChatTab/ChatControls/index.tsx
@@ -4,27 +4,12 @@ import { Icons } from '@onlook/ui/icons';
 import { Tooltip, TooltipContent, TooltipTrigger } from '@onlook/ui/tooltip';
 import { TooltipArrow } from '@radix-ui/react-tooltip';
 import { observer } from 'mobx-react-lite';
-import { useState } from 'react';
 
-const ChatControls = observer(() => {
+export const ChatControls = observer(() => {
     const editorEngine = useEditorEngine();
-    const [isHistoryOpen, setIsHistoryOpen] = useState(false);
-    const [showTooltip, setShowTooltip] = useState<boolean | undefined>(undefined);
-
     const handleNewChat = () => {
         editorEngine.chat.conversation.startNewConversation();
-        setIsHistoryOpen(false);
         editorEngine.chat.focusChatInput();
-    };
-
-    const handleHistoryOpenChange = (open: boolean) => {
-        setIsHistoryOpen(open);
-        // Force tooltip to close when dialog closes
-        if (!open) {
-            setShowTooltip(false);
-            // Reset tooltip state after a brief delay to allow hover to work again
-            setTimeout(() => setShowTooltip(undefined), 200);
-        }
     };
 
     return (
@@ -49,5 +34,3 @@ const ChatControls = observer(() => {
         </div>
     );
 });
-
-export default ChatControls;
