+++ apps/studio/src/routes/editor/EditPanel/ChatTab/ChatMessages/MessageContent/index.tsx
