+++ apps/studio/src/routes/editor/EditPanel/ChatTab/index.tsx
@@ -1,9 +1,9 @@
 import { observer } from 'mobx-react-lite';
 import { ChatInput } from './ChatInput';
-import ChatMessages from './ChatMessages';
+import { ChatMessages } from './ChatMessages';
 import { ErrorView } from './ErrorView';
 
-const ChatTab = observer(() => {
+export const ChatTab = observer(() => {
     return (
         <div className="w-full h-[calc(100vh-8.25rem)] flex flex-col justify-end gap-2">
             <ChatMessages />
@@ -12,5 +12,3 @@ const ChatTab = observer(() => {
         </div>
     );
 });
-
-export default ChatTab;
